file structure was fixed
