+++ lib/globals.dart
@@ -1,0 +1,6 @@
+double pos1 = 0;
+double pos2 = 0;
+bool flag = false;
+double deltaPos = 2.603;
+double globalPos = 0;
+bool scrollableFlag = true;
